+++ bench/fdt/windows/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,14 @@
+//
+//  Generated file. Do not edit.
+//
+
+// clang-format off
+
+#include "generated_plugin_registrant.h"
+
+#include <face_detection_tflite/face_detection_tflite_plugin.h>
+
+void RegisterPlugins(flutter::PluginRegistry* registry) {
+  FaceDetectionTflitePluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("FaceDetectionTflitePlugin"));
+}